remove null-ls. configure the lazy way
