--- after/plugin/autoclose.lua
@@ -1,1 +1,0 @@
-require("autoclose").setup({})
