--- after/plugin/colors.lua
@@ -1,9 +1,0 @@
-function ColorMyPencils(color)
-  color = color or "tokyonight-storm"
-  vim.cmd.colorscheme(color)
-
-  vim.api.nvim_set_hl(0, "Normal", { bg = "none" })
-  vim.api.nvim_set_hl(0, "NormalFloat", { bg = "none" })
-end
-
-ColorMyPencils()
