--- after/plugin/lua_ls.lua
@@ -1,10 +1,0 @@
-require'lspconfig'.lua_ls.setup {
-  settings = {
-    Lua = {
-      diagnostics = {
-        -- Get the language server to recognize the `vim` global
-        globals = {'vim'},
-      },
-    },
-  },
-}
