--- after/plugin/undotree.lua
@@ -1,1 +1,0 @@
-vim.keymap.set("n", "<leader>u", vim.cmd.UndotreeToggle)
