+++ lua/matt/init.lua
@@ -1,2 +1,35 @@
 require("matt.set")
 require("matt.remap")
+
+require("matt.lazy_init")
+
+local augroup = vim.api.nvim_create_augroup
+local MattGroup = augroup('matt, {}')
+local autocmd = vim.api.nvim_create_autocmd
+
+autocmd({ "BufWritePre" }, {
+  group = MattGroup,
+  pattern = "*",
+  command = [[%s/\s\+$//e]],
+})
+
+autocmd('LspAttach', {
+  group = MattGroup,
+  callback = function(e)
+    local opts = { buffer = e.buff }
+    vim.keymap.set("n", "gd", function() vim.lsp.buf.definition() end, opts)
+    vim.keymap.set("n", "K", function() vim.lsp.buf.hover() end, opts)
+    vim.keymap.set("n", "<leader>vws", function() vim.lsp.buf.workspace_symbol() end, opts)
+    vim.keymap.set("n", "<leader>vd", function() vim.diagnostic.open_float() end, opts)
+    vim.keymap.set("n", "[d", function() vim.diagnostic.goto_next() end, opts)
+    vim.keymap.set("n", "]d", function() vim.diagnostic.goto_prev() end, opts)
+    vim.keymap.set("n", "<leader>vca", function() vim.lsp.buf.code_action() end, opts)
+    vim.keymap.set("n", "<leader>vrr", function() vim.lsp.buf.references() end, opts)
+    vim.keymap.set("n", "<leader>vrn", function() vim.lsp.buf.rename() end, opts)
+    vim.keymap.set("i", "<C-h>", function() vim.lsp.buf.signature_help() end, opts)
+  end
+})
+
+vim.g.netrw_browse_split = 0
+vim.g.netrw_banner = 0
+vim.g.netrw_winsize = 25
