+++ lua/matt/lazy/autoclose.lua
@@ -1,0 +1,5 @@
+return {
+  "m4xshen/autoclose.nvim",
+  config = function()
+  end
+}
