+++ lua/matt/lazy/gitsigns.lua
@@ -1,0 +1,7 @@
+return {
+    "lewis6991/gitsigns.nvim",
+    config = function()
+      require("gitsigns").setup()
+    end
+    -- tag = 'release' -- To use the latest release (do not use this if you run Neovim nightly or dev builds!)
+}
