+++ lua/matt/lazy/prettier.lua
@@ -1,0 +1,5 @@
+return {
+  "MunifTanjim/prettier.nvim",
+  config = function()
+  end
+}
