+++ lua/matt/lazy/telescope.lua
@@ -1,0 +1,22 @@
+return {
+ "nvim-telescope/telescope.nvim",
+  version = "0.1.5",
+  dependencies = {  "nvim-lua/plenary.nvim" },
+  config = function()
+    local builtin = require('telescope.builtin')
+    vim.keymap.set('n', '<leader>pf', builtin.find_files, {})
+    vim.keymap.set('n', '<C-p>', builtin.git_files, {})
+    vim.keymap.set('n', '<leader>ps', function()
+      builtin.grep_string({ search = vim.fn.input("Grep > ") });
+    end)
+    vim.keymap.set('n', '<leader>pws', function()
+      local word = vim.fn.expand("<cword>")
+      builtin.grep_string({ search = word });
+    end)
+    vim.keymap.set('n', '<leader>pWs', function()
+      local word = vim.fn.expand("<cWORD>")
+      builtin.grep_string({ search = word });
+    end)
+    vim.keymap.let('n', '<leader>vh', builtin.help_tags, {})
+  end
+}
