+++ lua/matt/lazy/tmuxnavigator.lua
@@ -1,0 +1,4 @@
+return {
+    "christoomey/vim-tmux-navigator",
+    lazy = false,
+}
