--- lua/plugins/init.lua
@@ -1,79 +1,0 @@
-return {
-  { 'nvim-telescope/telescope.nvim', version = '0.1.5', dependencies = { { 'nvim-lua/plenary.nvim' } } },
-  {
-    "folke/tokyonight.nvim",
-    lazy = false,
-    priority = 1000,
-    opts = {},
-    config = function()
-      vim.cmd([[colorscheme tokyonight-night]])
-    end
-  },
-  {
-    'nvim-treesitter/nvim-treesitter',
-    build = ':TSUpdate',
-    config = function()
-      local configs = require('nvim-treesitter.configs')
-      configs.setup({
-        ensure_installed = { "c", "lua", "vim", "vimdoc", "query", "elixir", "heex", "javascript", "html", "typescript",
-          "python", "rust", "go" },
-        sync_install = false,
-        auto_install = true,
-        highlight = { enable = true, additional_vim_regex_highlighting = false },
-        indent = { enable = true },
-      })
-    end
-  },
-  { 'nvim-treesitter/playground' },
-  { 'ThePrimeagen/harpoon' },
-  { 'mbbill/undotree' },
-  { 'tpope/vim-fugitive' },
-  {
-    'VonHeikemen/lsp-zero.nvim',
-    dependencies = {
-
-      -- LSP Support
-      { 'neovim/nvim-lspconfig' },
-      { 'williamboman/mason.nvim' },
-      { 'williamboman/mason-lspconfig.nvim' },
-
-      -- Autocompletion
-      { 'hrsh7th/nvim-cmp' },
-      { 'hrsh7th/cmp-buffer' },
-      { 'hrsh7th/cmp-path' },
-      { 'saadparwaiz1/cmp_luasnip' },
-      { 'hrsh7th/cmp-nvim-lsp' },
-      { 'hrsh7th/cmp-nvim-lua' },
-
-      -- Snippets
-      { 'L3MON4D3/LuaSnip',                 version = "v2.*", build = "make install_jsregexp" },
-      { 'rafamadriz/friendly-snippets' },
-    }
-  },
-  { 'neovim/nvim-lspconfig' },
-  { 'jose-elias-alvarez/null-ls.nvim' },
-  { 'MunifTanjim/prettier.nvim' },
-  {
-    'nvim-tree/nvim-tree.lua',
-    dependencies = {
-      'nvim-tree/nvim-web-devicons', -- optional, for file icons
-    },
-    --tag = 'nightly' -- optional, updated every week. (see issue #1193)
-  },
-  {
-    'lewis6991/gitsigns.nvim',
-    config = function()
-      require('gitsigns').setup()
-    end
-    -- tag = 'release' -- To use the latest release (do not use this if you run Neovim nightly or dev builds!)
-  },
-  { 'm4xshen/autoclose.nvim' },
-  {
-    'nvim-lualine/lualine.nvim',
-    dependencies = { 'nvim-tree/nvim-web-devicons', lazy = true }
-  },
-  {
-    "christoomey/vim-tmux-navigator",
-    lazy = false,
-  }
-}
